turned json_parser into data_loader
