+++ app/models/data_loader.rb
@@ -1,0 +1,14 @@
+require 'json'
+
+class DataLoader
+  attr_reader :payload
+
+  def initialize(payload)
+    @payload = payload
+  end
+
+  def parsed_data
+    JSON.parse(payload)
+  end
+
+end
