--- app/models/json_parser.rb
@@ -1,14 +1,0 @@
-require 'json'
-
-class JsonParser
-  attr_reader :payload
-
-  def initialize(payload)
-    @payload = payload
-  end
-
-  def parsed_data
-    JSON.parse(payload)
-  end
-
-end
